+++ pandalone/xleash/io/_xlrd.py
@@ -14,21 +14,18 @@ Implements the *xlrd* backend of *xleash* that reads in-file Excel-spreadsheets.
 import datetime
 from distutils.version import LooseVersion
 import logging
-import os
-import sys
+from pandalone.xleash.io.backend import ABCBackend, ABCSheet, SheetId
 
 from future.moves.urllib import request
-from future.moves.urllib.parse import urlsplit
+from future.moves.urllib.parse import urlparse
 from xlrd import (xldate, XL_CELL_DATE, XL_CELL_EMPTY, XL_CELL_TEXT,
                   XL_CELL_BLANK, XL_CELL_ERROR, XL_CELL_BOOLEAN, XL_CELL_NUMBER)
 import xlrd
 
 import numpy as np
-import os.path as osp
 
-from .. import EmptyCaptureException, Coords
-from ... import utils
-from ._sheets import ABCSheet, SheetId
+from .. import EmptyCaptureException, Coords, io_backends
+from ... import utils, xlsutils
 
 
 log = logging.getLogger(__name__)
@@ -114,12 +111,10 @@ def _parse_cell(xcell, epoch1904=False):
                      (xcell.ctype, xcell.value))
 
 
-def _open_sheet_by_name_or_index(xlrd_book, wb_id, sheet_id, opts=None):
+def _open_sheet_by_name_or_index(xlrd_book, wb_id, sheet_id):
     """
     :param int or str or None sheet_id:
             If `None`, opens 1st sheet.
-    :param dict opts:
-            does nothing with them
     """
     if sheet_id is None:
         sheet_id = 0
@@ -136,40 +131,6 @@ def _open_sheet_by_name_or_index(xlrd_book, wb_id, sheet_id, opts=None):
             else:
                 xl_sh = xlrd_book.sheet_by_index(sheet_id)
     return XlrdSheet(xl_sh, wb_id)
-
-
-def open_sheet(wb_url, sheet_id, opts):
-    """
-    Opens the local or remote `wb_url` *xlrd* workbook wrapped as :class:`XlrdSheet`.
-    """
-    assert wb_url, (wb_url, sheet_id, opts)
-    ropts = opts.get('read', {})
-    if ropts:
-        ropts = ropts.copy()
-    if 'logfile' not in ropts:
-        level = logging.INFO if opts.get('verbose', None) else logging.DEBUG
-        ropts['logfile'] = utils.LoggerWriter(log, level)
-    parts = filename = urlsplit(wb_url)
-    if osp.isfile(wb_url) or not parts.scheme or parts.scheme == 'file':
-        fpath = osp.abspath(osp.expanduser(osp.expandvars(parts.path)))
-        log.info('Opening book %r...', fpath)
-        book = xlrd.open_workbook(fpath, **ropts)
-    else:
-        ropts.pop('on_demand', None)
-        http_opts = ropts.get('http_opts', {})
-        with request.urlopen('file://' + wb_url, **http_opts) as response:
-            log.info('Opening book %r...', filename)
-            book = xlrd.open_workbook(
-                filename, file_contents=response, **ropts)
-
-    return _open_sheet_by_name_or_index(book, wb_url, sheet_id, opts)
-
-
-def list_sheetnames(wb_id, opts=None):
-    if not opts:
-        opts = {}
-    # TODO: QnD list_sheetnames()!
-    return xlrd.open_workbook(wb_id, **opts).sheet_names()
 
 
 class XlrdSheet(ABCSheet):
@@ -197,11 +158,10 @@ class XlrdSheet(ABCSheet):
         return SheetId(self.book_fname or sh.book.filestr,
                        [sh.name, sh.number])
 
-    def open_sibling_sheet(self, sheet_id, opts=None):
+    def open_sibling_sheet(self, sheet_id):
         """Gets by-index only if `sheet_id` is `int`, otherwise tries both by name and index."""
         return _open_sheet_by_name_or_index(self._sheet.book,
-                                            self.book_fname, sheet_id,
-                                            opts)
+                                            self.book_fname, sheet_id)
 
     def list_sheetnames(self):
         return self._sheet.book.sheet_names()
@@ -243,3 +203,52 @@ class XlrdSheet(ABCSheet):
                 row.append(None)
 
         return table
+
+
+class XlrdBackend(ABCBackend):
+
+    def bid(self, wb_url):
+        if wb_url:
+            parts = urlparse(wb_url)
+            path = utils.urlpath2path(parts.path)
+            if xlsutils._xl_extensions.search(path):
+                return 100
+
+    def open_sheet(self, wb_url, sheet_id):
+        """
+        Opens the local or remote `wb_url` *xlrd* workbook wrapped as :class:`XlrdSheet`.
+        """
+        assert wb_url, (wb_url, sheet_id)
+        book = self._open_book(wb_url)
+
+        return _open_sheet_by_name_or_index(book, wb_url, sheet_id)
+
+    def list_sheetnames(self, wb_url):
+        # TODO: QnD list_sheetnames()!
+        book = self._open_book(wb_url)
+        return book.sheet_names()
+
+    def _open_book(self, url):
+        parts = filename = urlparse(url)
+        ropts = parts.params or {}
+        if 'logfile' not in ropts:
+            ropts['logfile'] = utils.LoggerWriter(log, logging.DEBUG)
+        if parts.scheme == 'file':
+            path = utils.urlpath2path(parts.path)
+            log.info('Opening book(%r)...', path)
+            book = xlrd.open_workbook(path, **ropts)
+        else:
+            ropts.pop('on_demand', None)
+            http_opts = ropts.get('http_opts', {})
+            with request.urlopen(url, **http_opts) as response:
+                log.info('Opening book(%r)...', filename)
+                book = xlrd.open_workbook(
+                    filename, file_contents=response, **ropts)
+
+        return book
+
+
+def load_as_xleash_plugin():
+    loaded = [be for be in io_backends if isinstance(be, XlrdBackend)]
+    if not loaded:
+        io_backends.insert(0, XlrdBackend())