+++ pandalone/xlsutils.py
@@ -23,33 +23,25 @@ __commit__ = ""
 
 log = logging.getLogger(__name__)
 
-# Probably Windows-only:
-#    TODO: Should become a class-method on xw.Workbook
-# when/if it works reliably, see github/xlwings #30.
+
+_xl_extensions = re.compile(r'\.xl((s[xm]?|t[xm]?)|w|m)\b', re.IGNORECASE)
 
 
-def get_active_workbook():
-    from win32com.client import dynamic
-    import xlwings as xw
-
-    com_app = dynamic.Dispatch('Excel.Application')
-    com_wb = com_app.ActiveWorkbook
-    wb = xw.Workbook(xl_workbook=com_wb)
-
-    return wb
+_xl_installed = None
 
 
-def get_Workbook(wrkb_fname):
-    """
-    :param str wrkb_fname: if missing return the active excel-workbook
-    """
-    import xlwings as xw
+def check_excell_installed():
+    """Checks once and returns `True` if Excel-app is installed in the system."""
+    global _xl_installed
+    if _xl_installed is None:
+        try:
+            from win32com.client import dynamic  # @UnresolvedImport
+            dynamic.Dispatch('Excel.Application')
+            _xl_installed = True
+        except Exception:  # pragma: no cover
+            _xl_installed = False
 
-    if wrkb_fname:
-        wb = xw.Workbook(wrkb_fname)
-    else:
-        wb = get_active_workbook()
-    return wb
+    return _xl_installed
 
 
 def _get_xl_vb_project(xl_wb):
@@ -210,8 +202,9 @@ def import_files_into_excel_workbook(infiles_wildcard, wrkb_fname=None, new_fnam
     :param str new_fname: a macro-enabled (`.xlsm`, `.xltm`) filepath for updated workbook,
                             `.xlsm` gets appended if current not a macro-enabled excel-format
     """
+    import xlwings as xw
 
-    wb = get_Workbook(wrkb_fname=None)
+    wb = xw.books.active
     xl_wb = wb.get_xl_workbook(wb)
     xl_vbp = _get_xl_vb_project(xl_wb)
     xl_vbcs = xl_vbp.VBComponents